turbo: hoist the visibility walk's frustum select and gate the player token lookup
CWorldObjList::UpdateVisibility (0x6838f0) rebuilt the active-frustum pointer
for every object it walked, an imul against the index global at 0xc7cfe0 plus
an add, because stock does that select per node inside its inlined
IsSphereCulled wrapper. The select is loop-invariant here, and the proof is in
the callees: both sphere tests are our own and read host globals only, and the
two stock delegates the walk reaches are leaves that never write 0xc7cfe0.
TSLink__ResolveNext 0x6876b0 has no store at all, and LinkedListNode__SetLinked
0x710b90 writes only through the list fields at its own receiver's +0x2c, +0x44
and +0x48. The pointer is materialized once before the head and reloaded into
ecx per call, which is what the register pressure allows: ebx, esi, edi and ebp
are all live across the test, so it lands in a frame slot rather than a
register. Two instructions and a load become one instruction and a load.

The same call also built a 16-byte stack copy of [center, radius] to hand over.
It does not need one: those four f32 are already contiguous at obj+0x5c,
TestSphere only reads them, and the only store between the load and the call is
the sort key at obj+0x78. The object's own address goes over instead, the way
the height-bucket test on the next line already passed it. The 16-byte store is
gone from the loop body and the lea targets the object.

transmog's ensure_bound opened with the client's unit-token lookup at 0x515970,
a name cascade through the object manager, and only then reached its cheap
guards. Two of those guards now come first. A receiver with no live descriptor
answers no GUID and cannot be any object, so it leaves without the lookup at
all. A receiver the binding already names leaves too: BOUND_GUID sits beside
BOUND_PLAYER so the fast path tests address and GUID together, which an address
alone cannot do, since an address is reusable once the object behind it is
freed. Reaching the lookup therefore means the binding does not name this
object, so the rebind below it is unconditional, which also closes the case
where the address came back holding a different player and the mirrors kept
describing the previous one. Verification is in-game: a character switch or a
reload must still rebind.

Two f64-to-integer casts in the command layer were the 64-bit form, which on
this target cannot stay in SSE and lowers to an fld/fisttp pair through the
frame with an unsigned fixup and a branch behind it. Both are bounded well
inside 32 bits. The frame cap's divisor is reached only on the else arm of
fps < 1.0 and fps > 500.0, so 1.0..=500.0 or a NaN, and both widths truncate
that range identically and answer a NaN with zero, which still divides by zero
as before; the edit is in set_interval, so it covers the foreground and the
background command at once. The timer command's delay and period are
milliseconds, where 32 bits is a 49-day ceiling, and narrowing them also keeps
the schedule's now + delay clear of the wrap a near-2^64 delay would otherwise
have fired immediately on. In the shipped image all three conversions are
maxsd/minsd/cvttsd2si with no x87 and no branch, and the fstp seams that
receive the client's own x87 result are untouched.

Rejected: deferring the opaque comparator's element reads past its flag test
(0x70ae10). The eager materialization is real, but the saving exists only on
the calls that leave before the key chain ties, and neither remedy is local.
read_draw_elem is shared with the key builder, which consumes every field of
every element and carries the sort weight, so a lazy read pessimizes the caller
that matters; deciding the first tier in the hooks layer duplicates the head of
a chain that lives in one place and whose operand order is load-bearing for the
unordered arm. The entry carries no samples in the capture. Why it stays eager
is written down at read_draw_elem instead.

Neither hooked entry touched here can be armed against the differential
harness, so the frustum hoist and the sphere pointer are verified from the
shipped codegen and from the stock callees' own bodies, and the coalescer and
the two commands are in-game observation.

+++ windows/turbo/src/win/unitxp.rs
@@ -379,8 +379,11 @@ fn timer_command(lua: &super::lua::LuaState) -> i32 {
                 return 1;
             };
             // the command truncates its millisecond doubles to integers, as
-            // the reference does
-            let id = timer::arm(delay as u64, period as u64, &script);
+            // the reference does, and takes them 32 bits wide: the ceiling
+            // that puts on a delay is 49 days, past anything a session can
+            // reach, and it keeps the schedule's `now + delay` clear of the
+            // wrap a near-2^64 delay would otherwise fire immediately on
+            let id = timer::arm(u64::from(delay as u32), u64::from(period as u32), &script);
             lua.push_number(f64::from(id));
             1
         }
